refactor: 토너먼트 진행 전용 파일을 _components/_apis/_hooks 등으로 이동
