+++ apps/web/src/app/tournament/[id]/_components/ProductCard.tsx
@@ -1,6 +1,6 @@
 import BaseImage from '@/components/common/base-image';
 
-import type { ProductT } from '../_types/tournament';
+import type { ProductT } from '../_common/_types/tournament';
 
 type ProductCardProps = ProductT & {
   isPicked?: boolean;
