+++ apps/web/src/app/tournament/[id]/page.tsx
@@ -3,11 +3,11 @@ import { notFound, redirect } from 'next/navigation';
 
 import { getQueryClient } from '@/utils/queryClient';
 
+import { postStartTournament } from './_apis/postStartTournament';
 import { getTournament } from './_common/_apis/getTournament';
-import { postStartTournament } from './_common/_apis/postStartTournament';
-import TournamentClient from './_common/_components/TournamentClient';
 import type { GetTournamentInProgressResponseT } from './_common/_types/tournamentResponse';
 import { parseTournamentId } from './_common/_utils/parseTournamentId';
+import TournamentClient from './_components/TournamentClient';
 
 type TournamentPageProps = {
   params: Promise<{ id: string }>;
